pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "A template for A2A agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
